fix: address ssh terminal ci review feedback
